requirements.txt
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Flask==3.1.3
flask-cors==4.0.0
gunicorn==25.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.0
Werkzeug==3.1.6
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-google-genai>=0.0.1
langchain-core>=0.1.0
langchain-pinecone>=0.1.0
langchain-huggingface>=0.0.1
pinecone-client>=2.2.0
